app/requirements.txt
fastapi
uvicorn
qdrant-client
langchain
langchain-openai
python-dotenv
qdrant-client

